# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alex production",
        "description": "Documentation for alex",
        "contact": {
            "email": "alex@gmail"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/brand": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get list of Brand",
                "description": "Returns list of Brand",
                "operationId": "indexBrand",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Insert a new Brand",
                "description": "Insert a new Brand",
                "operationId": "storeBrand",
                "requestBody": {
                    "description": "Brand to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Brand"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Brand created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Brand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/brand/{brand_id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Show a Brand from his Id",
                "description": "Show a Brand from his Id",
                "operationId": "showBrand",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Brand--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Brand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete a Brand",
                "description": "Delete a Brand",
                "operationId": "destroyBrand",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Brand--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Brand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update an existing Brand",
                "description": "Update an existing Brand",
                "operationId": "updateBrand",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Brand--id"
                    }
                ],
                "requestBody": {
                    "description": "Brand to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Brand"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Brand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Brand": {
                "title": "Brand",
                "description": "Brand model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Brand",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Brand",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Brand",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Brand",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Brand--id": {
                "name": "Brand_id",
                "in": "path",
                "description": "Id of Brand",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}